TestFrameGrabber:moved to src/devices and turn it to plugin
