+++ src/devices/TestFrameGrabber/testFrameGrabber.ini
@@ -1,0 +1,5 @@
+[plugin testFrameGrabber]
+type device
+name testFrameGrabber
+library yarp_testFrameGrabber
+wrapper grabber
